fix: build for non-automation
