+++ core/src/test/additional_disprove_scripts.rs
@@ -20,8 +20,7 @@ use crate::test::common::tx_utils::{
 use crate::test::common::{
     generate_withdrawal_transaction_and_signature, mine_once_after_in_mempool, run_single_deposit,
 };
-use crate::test::full_flow::get_tx_from_signed_txs_with_type;
-#[cfg(feature = "automation")]
+use crate::test::common::tx_utils::get_tx_from_signed_txs_with_type;
 use crate::utils::{FeePayingType, TxMetadata};
 use crate::{
     extended_rpc::ExtendedRpc,
